+++ angel_serialize_generator/lib/context.dart
@@ -59,6 +59,10 @@ class BuildContext {
   FieldElement get primaryKeyField =>
       fields.firstWhere((f) => f.name == primaryKeyName);
 
+  bool get importsPackageMeta {
+    return clazz.library.imports.any((i) => i.uri == 'package:meta/meta.dart');
+  }
+
   /// Get the aliased name (if one is defined) for a field.
   String resolveFieldName(String name) =>
       aliases.containsKey(name) ? aliases[name] : name;