+++ angel_serialize_generator/CHANGELOG.md
@@ -1,5 +1,8 @@
 # 2.0.7
 * Create unmodifiable Lists and Maps.
+* Support `@required` on fields.
+* Affix an `@immutable` annotation to classes, if
+`package:meta` is imported.
 
 # 2.0.6
 * Support for using `abstract` to create immutable model classes.